+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/AverageProfitCriterionTest.java
@@ -1,3 +1,26 @@
+/**
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2021 Ta4j Organization & respective
+ * authors (see AUTHORS)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package org.ta4j.core.analysis.criteria;
 
 import org.junit.Test;
@@ -16,17 +39,17 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.ta4j.core.TestUtils.assertNumEquals;
 
-public class AverageWinCriterionTest extends AbstractCriterionTest {
+public class AverageProfitCriterionTest extends AbstractCriterionTest {
 
-    public AverageWinCriterionTest(Function<Number, Num> numFunction) {
-        super((params) -> new AverageWinCriterion(), numFunction);
+    public AverageProfitCriterionTest(Function<Number, Num> numFunction) {
+        super((params) -> new AverageProfitCriterion(), numFunction);
     }
 
     @Test
     public void calculate() {
         BarSeries series = new MockBarSeries(numFunction, 100d, 95d, 102d, 105d, 97d, 113d, 115d, 120d, 110d, 90d);
-        TradingRecord tradingRecord = new BaseTradingRecord(
-                Trade.buyAt(0, series), Trade.sellAt(1, series), // losing -5
+        TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(0, series), Trade.sellAt(1, series), // losing
+                                                                                                             // -5
                 Trade.buyAt(2, series), Trade.sellAt(3, series), // winning +3
                 Trade.buyAt(4, series), Trade.sellAt(5, series), // winning +16
                 Trade.buyAt(6, series), Trade.sellAt(8, series) // losing -5
@@ -42,8 +65,7 @@ public class AverageWinCriterionTest extends AbstractCriterionTest {
     @Test
     public void calculateWithShortPositions() {
         BarSeries series = new MockBarSeries(numFunction, 100d, 95d, 102d, 105d, 97d, 113d);
-        TradingRecord tradingRecord = new BaseTradingRecord(
-                Trade.sellAt(0, series), Trade.buyAt(2, series),
+        TradingRecord tradingRecord = new BaseTradingRecord(Trade.sellAt(0, series), Trade.buyAt(2, series),
                 Trade.sellAt(3, series), Trade.buyAt(4, series));
         AnalysisCriterion avgWin = getCriterion();
